Fix /var unmounting bug
When a root partition has a separate /var partition, that /var
partition is mounted, added to the activated list, then unmounted.
The problem is that it is never removed from the activated list, thus
when all partitions are unmounted later on, an error occurs as YaST
tries to unmount it twice.