Fix crash in unit tests because of Yast::Convert.to_string
For some reason (maybe only in the tests?) Yast::Convert.to_string does
not convert nil to "", which is expected. To avoid tests failing because
of that, the variable assignation has been improved failing to "" in
case of nil.

Anyway, those old lines are going to be rewritten in incoming changes.